requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.30.2
accelerate==0.21.0
streamlit==1.28.0
pillow==10.0.0
numpy==1.24.3 